Update documentation on //... and //:all
The intent of this patch is to be a little more explicit about these
two patterns. It appears that there currently isn't a spot that
explicitly documents these patterns.

This patch was inspired by #13121. I'm hoping to submit a separate
patch to update the main FAQ (https://bazel.build/faq.html) with a
question along the lines of "How do I get bazel to build/test
everything?" and link to this new documentation here.

I also decided to update the examples for "Target names" and "Package
names" because the `...` syntax does have a special meaning. I thought
it best to avoid confusion for those cases.